moved Challenge and Layout to new files
